requirements.txt
fastapi
uvicorn
python-dotenv
PyPDF2
python-docx
httpx
aiofiles